+++ frontend/packages/types/src/index.ts
@@ -5,3 +5,6 @@ export * from './shared'
 export * from './system'
 export * from './test'
 export * from './queryParams'
+export * from './testKindStatistic'
+export * from './testStatusStatistic'
+export * from './systemScore'
